more api fixes for refactor
